refactor: 에셋 변경
