+++ src/pages/Home/components/HomeBackgroundLeft.tsx
@@ -1,5 +1,5 @@
 import homeLeftMountain from '@/assets/images/home-left-mountain.png';
-import homeLeftMountainDark from '@/assets/images/home-left-mountain-dark.png';
+import homeLeftMountainDark from '@/assets/images/home-left-mountain-dark.webp';
 
 import BackgroundImageWrapper from '@/components/BackgroundImageWrapper';
 import useThemeStore from '@/stores/themeStore';
